requirements.txt
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
funkify==0.4.4
idna==3.4
peewee==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.28.2
requires==0.10.3
service-identity==21.1.0
six==1.16.0
telebot==0.0.5
TelegramBotAPI==0.3.2
typing_extensions==4.5.0
urllib3==1.26.15
xtyping==0.6.2
